Allow to reproduce a CI run locally using libvirt
With the new job layout involving 3 nodes (CRC, Compute and an Ansible
controller), we have to provide people a way to reproduce such a job.

This patch extends the libvirt_manager role, as well as create a new
kind of "scenario" for reproducers. The intend there is to get multiple
"layouts" in order to mimic CI runs on a local environment.